refactor: Move callback state into SharedState
